+++ polyglot-test/tests/run-all-tests.rkt
@@ -1,0 +1,6 @@
+#lang racket/base
+
+(module+ test
+  (require rackunit/text-ui
+           "all-polyglot-tests.rkt")
+  (run-tests all-web-server-tests))